server/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
antlr4-python3-runtime==4.9.2
python-dotenv==1.0.0


